separated edit swtd from view, added the additional term implementations, filtered terms for the department of user
