+++ src/data/types.json
@@ -1,0 +1,7 @@
+{
+    "type" : [
+        "SEMESTER",
+        "MIDYEAR",
+        "SCHOOLYEAR"
+    ]
+}
