+++ src/pages/admin/Term.jsx
@@ -3,7 +3,8 @@ import Cookies from "js-cookie";
 import { useNavigate } from "react-router-dom";
 import { Form, Row, Col, Table } from "react-bootstrap";
 
-import { addTerm, getTerms } from "../../api/admin";
+import types from "../../data/types.json";
+import { addTerm, getTerms, deleteTerm } from "../../api/admin";
 import { isEmpty } from "../../common/validation/utils";
 import { useSwitch } from "../../hooks/useSwitch";
 import { useTrigger } from "../../hooks/useTrigger";
@@ -12,21 +13,29 @@ import SessionUserContext from "../../contexts/SessionUserContext";
 import BtnPrimary from "../../common/buttons/BtnPrimary";
 import ConfirmationModal from "../../common/modals/ConfirmationModal";
 import styles from "./style.module.css";
+import EditTermModal from "../../common/modals/EditTermModal";
 const Term = () => {
   const { user } = useContext(SessionUserContext);
   const token = Cookies.get("userToken");
   const navigate = useNavigate();
 
   const [showModal, openModal, closeModal] = useSwitch();
+  const [showEditModal, openEditModal, closeEditModal] = useSwitch();
+  const [showDeleteModal, openDeleteModal, closeDeleteModal] = useSwitch();
   const [showSuccess, triggerShowSuccess] = useTrigger(false);
+  const [showDeleteSuccess, triggerShowDeleteSuccess] = useTrigger(false);
+  const [showEditSuccess, triggerShowEditSuccess] = useTrigger(false);
   const [showError, triggerShowError] = useTrigger(false);
   const [errorMessage, setErrorMessage] = useState(null);
 
   const [terms, setTerms] = useState([]);
+  const [selectedTerm, setSelectedTerm] = useState(null);
+  const [selectedType, setSelectedType] = useState("");
   const [form, setForm] = useState({
     name: "",
     start_date: "",
     end_date: "",
+    type: "",
   });
 
   const fetchTerms = () => {
@@ -48,6 +57,7 @@ const Term = () => {
       ...form,
       [e.target.name]: e.target.value,
     });
+    if (e.target.name === "type") setSelectedType(e.target.value);
   };
 
   const clearForm = () => {
@@ -55,14 +65,17 @@ const Term = () => {
       name: "",
       start_date: "",
       end_date: "",
+      type: "",
     });
+    setSelectedType("");
   };
 
   const invalidFields = () => {
     const requiredFields = ["name", "start_date", "end_date"];
     return (
       requiredFields.some((field) => isEmpty(form[field])) ||
-      form.end_date < form.start_date
+      form.end_date < form.start_date ||
+      selectedType === ""
     );
   };
 
@@ -84,15 +97,37 @@ const Term = () => {
       },
       (response) => {
         triggerShowSuccess(3000);
+        setSelectedType("");
         fetchTerms();
         clearForm();
       },
       (error) => {
-        console.log(error.message);
         setErrorMessage(error.message);
         triggerShowError(3000);
       }
     );
+  };
+
+  const handleDelete = () => {
+    deleteTerm(
+      {
+        id: selectedTerm.id,
+        token: token,
+      },
+      (response) => {
+        triggerShowDeleteSuccess(4500);
+        fetchTerms();
+      },
+      (error) => {
+        setErrorMessage(error.message);
+        triggerShowError(3000);
+      }
+    );
+  };
+
+  const editSuccess = async () => {
+    triggerShowEditSuccess(3000);
+    await fetchTerms();
   };
 
   useEffect(() => {
@@ -118,63 +153,84 @@ const Term = () => {
         )}
 
         <Row>
-          <Form.Group as={Row} className="mb-3" controlId="inputName">
-            <Form.Label className={styles.formLabel} column sm="2">
-              Term
-            </Form.Label>
+          {/* Term Name */}
+          <Row>
+            <Form.Group as={Row} className="mb-3" controlId="inputName">
+              <Form.Label className={styles.formLabel} column sm="2">
+                Term Name
+              </Form.Label>
+              <Col sm="10">
+                <Form.Control
+                  className={styles.formBox}
+                  name="name"
+                  onChange={handleChange}
+                  value={form.name}
+                />
+              </Col>
+            </Form.Group>
+          </Row>
 
-            <Col sm="10">
-              <Form.Control
-                className={styles.formBox}
-                name="name"
-                onChange={handleChange}
-                value={form.name}
-              />
-            </Col>
-          </Form.Group>
-
-          <Row className="w-100">
-            <Col>
-              <Form.Group as={Row} className="mb-3" controlId="inputStartDate">
-                <Form.Label className={styles.formLabel} column sm="4">
-                  Start Date
-                </Form.Label>
-                <Col sm="8">
-                  <Form.Control
-                    type="date"
-                    className={styles.formBox}
-                    name="start_date"
+          {/* Term Type */}
+          <Row>
+            <Form.Group as={Row} className="mb-3" controlId="inputType">
+              <Form.Label className={`${styles.formLabel}`} column sm="2">
+                Term Type
+              </Form.Label>
+              <Col
+                className="d-flex justify-content-start align-items-center"
+                sm="10">
+                {types.type.map((item, index) => (
+                  <Form.Check
+                    key={index}
+                    type="radio"
+                    name="type"
+                    label={item}
+                    value={item}
                     onChange={handleChange}
-                    value={form.start_date}
+                    checked={selectedType === item}
+                    inline
                   />
-                </Col>
-              </Form.Group>
-            </Col>
+                ))}
+              </Col>
+            </Form.Group>
+          </Row>
 
-            <Col>
-              <Form.Group as={Row} className="mb-3" controlId="inputEndDate">
-                <Form.Label
-                  className={`${styles.formLabel} text-end`}
-                  column
-                  sm="4">
-                  End Date
-                </Form.Label>
-                <Col sm="8">
-                  <Form.Control
-                    type="date"
-                    min={form.start_date}
-                    className={styles.formBox}
-                    name="end_date"
-                    onChange={handleChange}
-                    value={form.end_date}
-                    isInvalid={form.end_date < form.start_date}
-                  />
-                  <Form.Control.Feedback type="invalid">
-                    End date must be after the start date.
-                  </Form.Control.Feedback>
-                </Col>
-              </Form.Group>
-            </Col>
+          {/* Dates */}
+          <Row>
+            <Form.Group as={Row} className="mb-3" controlId="inputStartDate">
+              <Form.Label className={`${styles.formLabel}`} column sm="2">
+                Start Date
+              </Form.Label>
+              <Col sm="3">
+                <Form.Control
+                  type="date"
+                  className={styles.formBox}
+                  name="start_date"
+                  onChange={handleChange}
+                  value={form.start_date}
+                />
+              </Col>
+            </Form.Group>
+
+            <Form.Group as={Row} controlId="inputEndDate">
+              <Form.Label className={`${styles.formLabel}`} column sm="2">
+                End Date
+              </Form.Label>
+              <Col sm="3">
+                <Form.Control
+                  type="date"
+                  min={form.start_date}
+                  className={styles.formBox}
+                  name="end_date"
+                  onChange={handleChange}
+                  value={form.end_date}
+                  isInvalid={form.end_date < form.start_date}
+                />
+                <Form.Control.Feedback type="invalid">
+                  End date must be after the start date.
+                </Form.Control.Feedback>
+              </Col>
+            </Form.Group>
           </Row>
         </Row>
 
@@ -197,6 +253,18 @@ const Term = () => {
         </Row>
       </Form>
       <hr />
+
+      {showEditSuccess && (
+        <div className={`${styles.form} alert alert-success mb-3`} role="alert">
+          Term details updated successfully.
+        </div>
+      )}
+
+      {showDeleteSuccess && (
+        <div className={`${styles.form} alert alert-success mb-3`} role="alert">
+          Term deleted.
+        </div>
+      )}
       <Row className={`${styles.table}  w-100`}>
         {terms.length === 0 ? (
           <Col className="text-center">No terms added yet.</Col>
@@ -206,8 +274,10 @@ const Term = () => {
               <tr>
                 <th>ID</th>
                 <th>Name</th>
+                <th>Type</th>
                 <th>Start Date</th>
                 <th>End Date</th>
+                <th className="text-center">Actions</th>
               </tr>
             </thead>
             <tbody>
@@ -216,11 +286,41 @@ const Term = () => {
                   <tr key={term.id}>
                     <td>{term.id}</td>
                     <td>{term.name}</td>
+                    <td>{term.type}</td>
                     <td>{term.start_date}</td>
                     <td>{term.end_date}</td>
+                    <td className="text-center">
+                      <i
+                        className={`${styles.icon} fa-solid fa-pen-to-square text-dark me-3`}
+                        onClick={() => {
+                          openEditModal();
+                          setSelectedTerm(term);
+                        }}></i>
+                      <i
+                        className={`${styles.icon} fa-solid fa-trash-can text-danger`}
+                        onClick={() => {
+                          openDeleteModal();
+                          setSelectedTerm(term);
+                        }}></i>
+                    </td>
                   </tr>
                 ))}
             </tbody>
+            <EditTermModal
+              show={showEditModal}
+              onHide={closeEditModal}
+              data={selectedTerm}
+              editSuccess={editSuccess}
+            />
+            <ConfirmationModal
+              show={showDeleteModal}
+              onHide={closeDeleteModal}
+              onConfirm={handleDelete}
+              header={"Delete Term"}
+              message={
+                "Are you sure about deleting this term? It must not have any SWTD submissions linked to it."
+              }
+            />
           </Table>
         )}
       </Row>
