+++ src/pages/admin/style.module.css
@@ -44,3 +44,7 @@
   border-radius: 7px;
   border: none;
 }
+
+.icon {
+  cursor: pointer;
+}
